+++ Search/degrees/small/movies.csv
@@ -1,0 +1,6 @@
+id,title,year
+112384,"Apollo 13",1995
+104257,"A Few Good Men",1992
+109830,"Forrest Gump",1994
+93779,"The Princess Bride",1987
+95953,"Rain Man",1988
