+++ Search/degrees/small/people.csv
@@ -1,0 +1,17 @@
+id,name,birth
+102,"Kevin Bacon",1958
+129,"Tom Cruise",1962
+144,"Cary Elwes",1962
+158,"Tom Hanks",1956
+1597,"Mandy Patinkin",1952
+163,"Dustin Hoffman",1937
+1697,"Chris Sarandon",1942
+193,"Demi Moore",1962
+197,"Jack Nicholson",1937
+200,"Bill Paxton",1955
+398,"Sally Field",1946
+420,"Valeria Golino",1965
+596520,"Gerald R. Molen",1935
+641,"Gary Sinise",1955
+705,"Robin Wright",1966
+914612,"Emma Watson",1990
