+++ Search/degrees/small/stars.csv
@@ -1,0 +1,21 @@
+person_id,movie_id
+102,104257
+102,112384
+129,104257
+129,95953
+144,93779
+158,109830
+158,112384
+1597,93779
+163,95953
+1697,93779
+193,104257
+197,104257
+200,112384
+398,109830
+420,95953
+596520,95953
+641,109830
+641,112384
+705,109830
+705,93779
